feat: extract SteeringRepository from AppPersistence
Move steering corrections, audit log, token usage, usage summary,
and full-text search into a dedicated SteeringRepository module.

- Create SteeringRepository.ts with factory function pattern
- Extract 5 correction methods + logAudit + insertTokenUsage + getUsageSummary
- Implement missing search() via FTS5 (kanban_cards, documents, decisions)
- Implement missing getAuditLog() with AuditLogFilters type
- Add backupDatabase() to AppPersistenceService interface
- AppPersistence now delegates all extracted methods to the repository
- Re-export AuditLogFilters, AuditLogEntry, SearchResult types
